have flights showing up just need to put in loop
