+++ package.json
@@ -23,6 +23,7 @@
     "express-handlebars": "^5.2.1",
     "mysql": "^2.18.1",
     "mysql2": "^2.2.5",
+    "node-fetch": "^2.6.1",
     "sequelize": "^6.5.0",
     "sequelizer": "^1.1.4"
   }
